requirements.txt
pytest==7.4.4
python-dotenv==1.0.1
selenium==4.16.0
allure-pytest==2.13.2
